requirements.txt
PyPDF2
langchain
langchain_google_genai
google.generativeai

langchain_google_genai

python-dotenv